pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turbo"
version = "1.0.0"
description = "AI-powered local project management and development platform"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Turbo Development Team", email = "dev@turbo.local"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.10"
dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.28.0",
    "greenlet>=3.0.0",

    # Web UI
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.17.0",
    "jinja2>=3.1.0",

    # CLI
    "typer[all]>=0.9.0",
    "rich>=13.6.0",
    "click>=8.1.0",

    # Utilities
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "python-dateutil>=2.8.0",
    "toml>=0.10.0",
    "email-validator>=2.0.0",

    # Export/Import
    "python-docx>=1.1.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
    "PyYAML>=6.0",
    "weasyprint>=62.0",  # PDF generation from HTML

    # Content Extraction (Literature)
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.0",

    # Resume Parsing
    "pdfplumber>=0.11.0",
    "anthropic>=0.40.0",
    "rapidfuzz>=3.0.0",  # Fuzzy string matching for deduplication

    # Audio Transcription (Local)
    "faster-whisper>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "pyannote.audio>=3.1.0",  # Speaker diarization
    "torch>=2.0.0",  # Required by pyannote
    "torchaudio>=2.0.0",  # Audio processing for torch

    # Security
    "passlib[bcrypt]>=1.7.0",
    "python-jose[cryptography]>=3.3.0",

    # Terminal
    "ptyprocess>=0.7.0",
    "websockets>=12.0",

    # Knowledge Graph
    "neo4j>=5.14.0",
    "sentence-transformers>=2.2.0",

    # LLM Integration
    "ollama>=0.4.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.3.0",
    "httpx>=0.25.0",  # For FastAPI testing

    # Code Quality
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",

    # Type Stubs
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.8.0",
    "types-toml>=0.10.0",

    # Development Tools
    "ipython>=8.16.0",
    "jupyter>=1.0.0",
    "watchdog>=3.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.6.0",
    "pymdown-extensions>=10.16.0",
]
mcp = [
    # Model Context Protocol for Claude Code integration
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/username/turbo"
Documentation = "https://turbo.readthedocs.io"
Repository = "https://github.com/username/turbo"
Issues = "https://github.com/username/turbo/issues"

[project.scripts]
turbo = "turbo.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["turbo*"]
exclude = ["tests*", "docs*"]

# Black Configuration
[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

# Ruff Configuration
[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "S",   # flake8-bandit (security)
    "N",   # pep8-naming
    "ERA", # eradicate (commented code)
    "PL",  # pylint
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "S101",   # use of assert detected
    "PLR0913", # too many arguments to function call
    "PLR0915", # too many statements
    "PLR2004", # magic value used in comparison
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

[tool.ruff.per-file-ignores]
"tests/**/*" = ["S101", "ARG", "PLR2004"]
"scripts/**/*" = ["S101", "PLR2004"]

[tool.ruff.isort]
known-first-party = ["turbo"]
force-sort-within-sections = true

[tool.ruff.flake8-type-checking]
strict = true

# MyPy Configuration
[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
no_implicit_reexport = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "scripts.*"
]
disallow_untyped_defs = false
disallow_incomplete_defs = false

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=turbo",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=85",
    "-ra",
    "--tb=short"
]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning:streamlit.*",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "claude: marks tests that require Claude integration",
]
asyncio_mode = "auto"

# Coverage Configuration
[tool.coverage.run]
source = ["turbo"]
branch = true
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/scripts/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "\\.\\.\\.",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# Bandit Security Configuration
[tool.bandit]
exclude_dirs = ["tests", "scripts"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection_process